requirements.txt
graphviz==0.8.2
h5py==2.7.1
Keras==2.1.3
matplotlib==2.1.2
numpy==1.14.0
pathos==0.2.1
Pillow==5.0.0
pydot==1.2.4
scipy==1.0.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
tqdm==4.19.5
